Add pspec evaluator and parser
This commit adds a pspec evaluator that uses the puppet parser to parse
the expectations. Internally, the parser uses a PN parser to parse
such expressions, thereby removing the need to adjust for proper line
breaks and indents.

The commit also changes some of the PN call names to avoid conflict such
as '()' and '$'

foo