Make all issues use named and unformatted parameters
Prior to this commit, all issues used pre-formatted and positional
arguments. The formatting (adding article, sometimes in uppercase) was
handled by the caller. This was an interrim quick and dirty solution
which has now been replaced by a proper implementation. The issue takes
named arguments and adding articles and such is done in the issue.