Change the Data representation of a PN Map so that order is preserved
This commit changes the `Data` representation of the PN Map into a
one element Hash where the key is '#' and the value is an array with
an even number of elements where each evenly positioned element is a
key for the next element in the list.

The reason for this change is that a standard Hash is represented as
an object in JSON. The JSON object does not preserve the order of its
entries.