Resolve Object type declaration variants early
Prior to this commit, the Parser would emit different variants of
an Object data type declation depending on which variant that was
used. This commit ensures that the AST will use the canonical
representation with an AccessExpression in the form Object[{decl}]
regardless of the actual declaration.

The change is motivaded by that this is how the Ruby implementation
works. It also simplifies the validator and evaluator implementation
of the Object data type.