Fix issue with sequences of named expression
A NamedExpression that had a NamedExpression as its left hand side
operand must resolve that operand to a CallMethodExpression. That didn't
happen prior to this commit. So x.y worked fine but x.y.z did not. This
commit adds the necessary logic to transform the LHS into a call.