(GEN-70) Add conditional handling of workflow to the Puppet DSL parser.
This commit adds the keyword 'workflow' to the Puppet DSL. The keyword
is only recogniced when the option PARSER_WORKFLOW_ENABLED is set.

A workflow consists of a set of activities where each activity can be
an action, resource, or a workflow. The workflow is introduced as part
of the Genesis project and that is also where the specification can be
found.